add possibility to use `embedded-hal-bus`
this isn't particularly nice, but it avoids re-writing the complete
library to kick out `Driver: I2c + DelayNs` which permeates all of the
code.

a better name might be required for this!

note that depending on the HAL you might have the possibility to clone
the delay or instantiate a generic new one every time (e.g. using
`embassy_time::Delay`).

resolves #16